# optimized dot by 6x in some cases