used sum instead of einsum

[cpu] Shape=(512, 64, 64) Dtype=torch.float32
  Einsum:  1021.676 ms
  Sum:     181.976 ms
  Speedup: 5.61x
  Max diff: 9.536743e-06

[cpu] Shape=(128, 128, 128) Dtype=torch.complex64
  Einsum:  1275.972 ms
  Sum:     767.676 ms
  Speedup: 1.66x
  Max diff: 1.417739e-05

[cuda] Shape=(512, 64, 64) Dtype=torch.float32
  Einsum:  71.882 ms
  Sum:     14.387 ms
  Speedup: 5.00x
  Max diff: 5.722046e-06

[cuda] Shape=(128, 128, 128) Dtype=torch.complex64
  Einsum:  62.479 ms
  Sum:     50.555 ms
  Speedup: 1.24x
  Max diff: 6.031566e-06

heres the benchmarking and validation script 
https://colab.research.google.com/drive/1FNTqdX3f9ub3Jm68Vnw1zmbswEpxezT_?usp=sharing